move out external-services test
