+++ enterprise/internal/authz/authz.go
@@ -191,3 +191,11 @@ func PermissionSyncingDisabled() bool {
 		licensing.Check(licensing.FeatureACLs) != nil ||
 		conf.Get().DisableAutoCodeHostSyncs
 }
+
+var ValidateExternalServiceConfig = database.MakeValidateExternalServiceConfigFunc(
+	[]func(*types.GitHubConnection) error{github.ValidateAuthz},
+	[]func(*schema.GitLabConnection, []schema.AuthProviders) error{gitlab.ValidateAuthz},
+	[]func(*schema.BitbucketServerConnection) error{bitbucketserver.ValidateAuthz},
+	[]func(*schema.PerforceConnection) error{perforce.ValidateAuthz},
+	[]func(*schema.AzureDevOpsConnection) error{func(_ *schema.AzureDevOpsConnection) error { return nil }},
+) // TODO: @varsanojidan switch this with actual authz once its implemented.
